+++ libbeat/reader/message.go
