+++ libbeat/reader/multiline/multiline_test.go
