+++ libbeat/reader/readfile/encoding/encoding.go
