+++ libbeat/reader/readfile/encoding/mixed.go
