+++ libbeat/reader/readfile/encoding/utf16.go
