+++ libbeat/reader/readfile/line_test.go
