+++ libbeat/reader/readfile/strip_newline_test.go
