+++ libbeat/reader/readjson/docker_json_test.go
