+++ libbeat/reader/readjson/json_test.go
